fix(routing): move hotel directory to src/app to resolve Next.js conflicting router paths
