+++ src/app/hotel/[id]/escrow/create/page.tsx
@@ -1,44 +1,50 @@
-"use client";
+'use client';
 
-import Link from "next/link";
-import { useParams } from "next/navigation";
-import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
+import { HotelHeader } from '@/components/hotel';
+import Link from 'next/link';
+import { use } from 'react';
 
-/**
- * Shown when escrow is still pending signature / deployment.
- * Users are redirected here from the escrow detail route until signing completes.
- */
-export default function HotelEscrowCreatePage() {
-  const params = useParams<{ id: string }>();
-  const hotelId = params.id;
+export default function HotelEscrowCreatePage({
+  params,
+}: {
+  params: Promise<{ id: string }>;
+}) {
+  const resolvedParams = use(params);
 
   return (
-    <div className="min-h-screen bg-gradient-to-b from-slate-50 to-slate-100 dark:from-slate-950 dark:to-slate-900 flex items-center justify-center p-4">
-      <Card className="w-full max-w-lg">
-        <CardHeader>
-          <CardTitle>Complete escrow signing</CardTitle>
-          <CardDescription>
-            Hotel ID: {hotelId}. Signing has not finished yet. Continue the create
-            flow to deploy your escrow, then return to the detail page to track
-            status.
-          </CardDescription>
-        </CardHeader>
-        <CardContent className="flex flex-col gap-3 sm:flex-row">
-          <Button asChild className="w-full sm:w-auto">
-            <Link href={`/dashboard/hotel/create-escrow`}>Create escrow</Link>
-          </Button>
-          <Button variant="outline" asChild className="w-full sm:w-auto">
-            <Link href={`/dashboard/hotel`}>Hotel dashboard</Link>
-          </Button>
-        </CardContent>
-      </Card>
+    <div className="min-h-screen bg-[#faf7f3]">
+      <HotelHeader />
+      <main className="mx-auto max-w-[860px] px-6 py-16">
+        <div className="rounded-[18px] border border-[#e7ddd5] bg-white p-8 shadow-sm">
+          <span className="inline-flex rounded-full bg-[#fff1e7] px-3 py-1 text-xs font-semibold uppercase tracking-[0.08em] text-[#ff6a00]">
+            Escrow flow
+          </span>
+          <h1 className="mt-5 text-[34px] font-semibold tracking-[-0.04em] text-[#1d1d1d]">
+            Create escrow for apartment {resolvedParams.id}
+          </h1>
+          <p className="mt-4 max-w-[620px] text-base leading-7 text-[#666666]">
+            This route is wired so the BOOK button from the apartment detail
+            page lands on a valid escrow creation screen instead of a 404. The
+            full escrow form can be layered onto this route in the follow-up
+            issue.
+          </p>
+
+          <div className="mt-8 flex flex-wrap gap-4">
+            <Link
+              href={`/hotel/${resolvedParams.id}`}
+              className="rounded-[10px] border border-[#d7cdc4] px-5 py-3 text-sm font-medium text-[#282828]"
+            >
+              Back to apartment
+            </Link>
+            <Link
+              href="/hotel"
+              className="rounded-[10px] bg-[#ff6a00] px-5 py-3 text-sm font-semibold text-white"
+            >
+              Browse more apartments
+            </Link>
+          </div>
+        </div>
+      </main>
     </div>
   );
 }
