+++ src/app/hotel/[id]/page.tsx
