+++ src/app/hotel/page.tsx
